# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API for travel agency",
        "description": "API for travel agency",
        "contact": {
            "email": "example@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "API for travel agency server"
        }
    ],
    "paths": {
...
        "/tours": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Get list of tours",
                "description": "Returns list of tours",
                "operationId": "getTours",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "app_id": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tours"
                ],
                "operationId": "ab7a9763f9aa040271e38eae5067e34b",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string"
                                            },
                                            "place_id": {
                                                "type": "integer"
                                            },
                                            "discount_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "SkyRun Adventure",
                                    "price": "399",
                                    "description": "SkyRun Adventure is an exhilarating tour that takes you high above the ground, giving you a bird's-eye view of the stunning scenery below",
                                    "image": "lW9pR5ma7o5NuUemvwbHlIsKym64gXqAp6moLt8z.png",
                                    "place_id": "5",
                                    "discount_id": "6"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "SkyRun Adventure"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "399"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "SkyRun Adventure is an exhilarating tour that takes you high above the ground, giving you a bird's-eye view of the stunning scenery below"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "images/lW9pR5ma7o5NuUemvwbHlIsKym64gXqAp6moLt8z.png"
                                        },
                                        "place_id": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "discount_id": {
                                            "type": "integer",
                                            "example": "6"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/tours/{id}": {
            "get": {
                "tags": [
                    "Tours"
                ],
                "summary": "Get tour information",
                "description": "Returns tour data",
                "operationId": "getTourById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Tour id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tours"
                ],
                "operationId": "e066b1d619d620065efa6ba2f5d86df3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "price": {
                                                "type": "integer"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "image": {
                                                "type": "string"
                                            },
                                            "place_id": {
                                                "type": "integer"
                                            },
                                            "discount_id": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "SkyRun Adventure",
                                    "price": "399",
                                    "description": "SkyRun Adventure is an exhilarating tour that takes you high above the ground, giving you a bird's-eye view of the stunning scenery below",
                                    "image": "lW9pR5ma7o5NuUemvwbHlIsKym64gXqAp6moLt8z.png",
                                    "place_id": "5",
                                    "discount_id": "6"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "SkyRun Adventure"
                                        },
                                        "price": {
                                            "type": "integer",
                                            "example": "399"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "SkyRun Adventure is an exhilarating tour that takes you high above the ground, giving you a bird's-eye view of the stunning scenery below"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "images/lW9pR5ma7o5NuUemvwbHlIsKym64gXqAp6moLt8z.png"
                                        },
                                        "place_id": {
                                            "type": "integer",
                                            "example": "5"
                                        },
                                        "discount_id": {
                                            "type": "integer",
                                            "example": "6"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tours"
                ],
                "operationId": "f3749acb3457cf602b904a571e02eee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "delete tour success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}